API

FlightLib

Dynamics

Flight.FlightLib.Dynamics.FrameTransform — Type

Specify a reference frame $c$ with respect to another $b$.

$c$ is defined by:

$r_{bc}^{b}$: Position vector from $b$'s origin $O_b$ to $c$'s origin $O_c$, projected on $b$'s axes $Ɛ_b$

$q^b_c$: Rotation quaternion from $b$'s axes $Ɛ_b$ to $c$'s axes $Ɛ_c$

FrameTransform objects are callable; the syntax t_bc(x_c) is equivalent to translate(t_bc, x_c).

Flight.FlightLib.Dynamics.translate — Method

translate(t_bc::FrameTransform, r_cP_c::AbstractVector{<:Real})

Translate a 3D position vector from one reference frame to another.

Given the FrameTransform t_bc from frame $b$ to frame $c$ and the position vector of some point $P$ in $c$ ($r_{cP}^{c}$), return the position vector of $P$ in $b$ ($r_{bP}^{b}$).

Base.adjoint — Method

Base.:adjoint(t_bc::FrameTransform)

Given t_bc, return its reciprocal t_cb.

Base.:∘ — Method

Base.:∘(t_bc::FrameTransform, t_cd::FrameTransform)

Concatenate FrameTransforms t_bc and t_cd and return the resulting t_bd.

Flight.FlightLib.Dynamics.Wrench — Type

Force and torque combination defined in a specific reference frame.

A Wrench is defined in frame $c$ when it is applied at $c$'s origin $O_c$ and projected on its axes $Ɛ_c$

Flight.FlightLib.Dynamics.FrameTransform

Flight.FlightLib.Dynamics.Wrench

Base.:∘

Base.adjoint

Flight.FlightLib.Dynamics.translate